+++ StarRender/src/Font.cpp
@@ -1,4 +1,5 @@
 #include "Font.h"
+#include <glad/glad.h>
 #include <ft2build.h>
 #include FT_FREETYPE_H
 #include <glm/gtc/matrix_transform.hpp>
@@ -35,7 +36,7 @@ void Font::Init(std::string filePath, int pSize) {
 			texture,
 			glm::ivec2(fontFace->glyph->bitmap.width, fontFace->glyph->bitmap.rows),
 			glm::ivec2(fontFace->glyph->bitmap_left, fontFace->glyph->bitmap_top),
-			fontFace->glyph->advance.x
+			static_cast<uint32_t>(fontFace->glyph->advance.x)
 		};
 		mFontCharacters.insert(std::pair<GLchar, FontCharacter>(c, character));
 	}
